util/ctx: add guarded context manager
This adds a new Class `util.ctx.GuardedCM`. This implements a simple
context-manager that allows any class that inherits `GuardedCM` to
define `guarded_enter()` and `guarded_exit()` rather than the standard
`__enter__()` and `__exit__()` methods. Unlike the standard methods,
this variant guarantees that `guarded_exit()` is called even if the
enter operation raises an exception.

Additionally, the `GuardedCM` class provides a builtin exit-stack. With
this, you can open more context-managers from within `guarded_enter()`
as if you would stack all of them in a `contextlib.ExitStack`.
Similarly, this guarantees that the exit-stack is cleaned up properly
whenever anything fails.

IOW, this context-manager guarantees that whenever you do something in
`guarded_enter()`, you will have the chance to clean it up in
`guarded_exit()`.