test: add FedMir manifest
Add a F32 manifest with a very simple RPM stage. This manifest is meant
to contain all packages that we might require in our test-suite. The
idea is to include them in this manifest, but use `localhost:8071` as
base-url. We then build this manifest early in the CI, and thus populate
our cache with all required packages fetched from the local FedMir.

Rather than adjusting all our manifests to use `localhost:8071`, we keep
them with a working global mirror (usually kernel.org in our case). By
building the new `fedora-fedmir.json` first, we make sure the store
contains all sources prefetched, so later builds do not have to pull in
all the packages again.

This only adds the (MPP-)manifest for `fedora-fedmir.json`. It is not
used, yet.